Prepare app for Play release and add safe migrations

+++ app/src/main/java/com/epatay/digitalwallet/data/InvestmentItem.kt
@@ -8,7 +8,7 @@ data class InvestmentItem(
     @PrimaryKey(autoGenerate = true)
     val id: Int = 0,               // Otomatik artan takip numarası
     val assetName: String,         // Örn: "Altın", "USD", "Euro"
-    val amount: Int,            // Ne kadar alındı? (Örn: 10)
+    val amount: Double,         // Ne kadar alındı? (Örn: 10.5)
     val buyPrice: Double,          // Alındığı anki fiyat/kur (Örn: 32.50)
     val buyDate: String            // Alış tarihi ve saati
 )
